Prevent most flickering while rendering
Instead of moving the cursor up and clearing each line before rendering the output, move to the top first, print each line, and then clear the rest of the line to erase old output. This avoids the case where all lines are empty for a few milliseconds as the sequences are being parsed.