Reuse StringBuilder and avoid intermediate String
Reuse the StringBuilder because its backing character array is likely to be sized appropriately for the next frame. We incur a penalty of writing null chars into the whole array, but we would pay that anyway with a new allocation.

Additionally, use a CharBuffer/CharEncoder/ByteBuffer combination to avoid needing to render to an intermediate String before conversion to bytes. This should avoid quite a few copies since we can write a subset of the resulting ByteBuffer rather than needing to copy it first to a perfectly-sized byte array.