Clear line before rendering to avoid deleting final character (#706)
When a row writes into the last column, the cursor stays in that position. Clearing the remaining line from that position has the potential to delete the final character. Instead, clear the line before rendering it. This may cause partial row flickering for terminals which do not support synchronized rendering, but that's an okay trade-off.

Additionally, use more efficient ANSI to move the cursor up and to clear the remaining lines.